Use args.job instead of sys.argv[1]
The call to loop() uses sys.argv[1], but should use args.job instead.  When I tried running it like "tail.py -a MYADDRESS -u user -p pass JOBID", sys.argv[1] is set to "-a", so that becomes the job_id in loop(), which makes the script only able to pull the first page of output since subsequent calls have a bad job_id.